NDS: Progress towards booting games Fixed LDM/STM edge cases on ARM and Thumb, of which LDMIA's on ARM7 prevented BIOS 7 from properly sending the commands to retrieve the secure area Properly re-encrypt the secure area to be decrypted by the BIOS Add cartridge DMAs Add cartridge enable/disable in AUXSPICNT Fix a DMA bug where writing to DMACNT with the enable bit set to 1 would trigger an immediate mode DMA even if the bit was already set previously (immediate mode DMAs are edge-triggered on hardware) Add identification mode to firmware SPI flash Add POWCNT1/POWCNT2 registers Set shared memory mode before direct booting since some games load their ARM7 binaries to shared memory

+++ src/emulator/MemoryUtil.cs
@@ -12,16 +12,15 @@ namespace OptimeGBA
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public static ulong GetUlong(byte[] arr, uint addr)
         {
-            return (ulong)(
-                    (arr[addr + 0] << 0) |
-                    (arr[addr + 1] << 8) |
-                    (arr[addr + 2] << 16) |
-                    (arr[addr + 3] << 24) |
-                    (arr[addr + 4] << 32) |
-                    (arr[addr + 5] << 40) |
-                    (arr[addr + 6] << 48) |
-                    (arr[addr + 7] << 56)
-                );
+            return
+                ((ulong)arr[addr + 0] << 0) |
+                ((ulong)arr[addr + 1] << 8) |
+                ((ulong)arr[addr + 2] << 16) |
+                ((ulong)arr[addr + 3] << 24) |
+                ((ulong)arr[addr + 4] << 32) |
+                ((ulong)arr[addr + 5] << 40) |
+                ((ulong)arr[addr + 6] << 48) |
+                ((ulong)arr[addr + 7] << 56);
         }
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
